Move FieldTypes utility to where it belongs.
