+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/message/fieldtype/FieldTypes.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.gradle.compiler.util;
+package io.spine.gradle.compiler.message.fieldtype;
 
 import com.google.protobuf.DescriptorProtos.FieldDescriptorProto;
